Move atmosphere and barycentric mask shifting code to separate file atmosphere.py
